pyproject.toml
[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aiida-gui"
dynamic = ["version"]  # read from aiida_gui/__init__.py
description = "Design flexible node-based workflow for AiiDA calculation."
authors = [{name = "Xing Wang", email = "xingwang1991@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 1 - Planning",
    "Framework :: AiiDA",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering"
]
keywords = ["aiida", "workflows"]
requires-python = ">=3.9"
dependencies = [
    "numpy~=1.21",
    "scipy",
    "aiida-core>=2.3",
    "aiida_workgraph",
    "cloudpickle",
    "fastapi",
    "uvicorn",
    "pydantic_settings",
    "weas_widget",
]

[project.urls]
Documentation = "https://aiida-gui.readthedocs.io"
Source = "https://github.com/aiidateam/aiida-gui"

[project.optional-dependencies]
docs = [
    "sphinx_rtd_theme",
    "sphinx~=7.2",
    "sphinx-copybutton~=0.5.0",
    "sphinx-design~=0.5.0",
    "sphinx-notfound-page~=1.0",
    "sphinxext-rediraffe~=0.2.4",
    "sphinx-intl~=2.1.0",
    "sphinx-gallery",
    "myst-nb~=1.0.0",
    "nbsphinx",
]
pre-commit = [
    "pre-commit~=2.2",
    "pylint~=2.17.4",
]
tests = [
    'pgtest~=1.3',
    "pytest~=7.0",
    "pytest-cov~=2.7,<2.11",
    "playwright",
    "httpx",
]

[project.scripts]
workgraph_web_ui = "aiida_gui.cmd_web:web"

[project.entry-points."workgraph.cmdline"]
"web" = "aiida_gui.cmd_web:web"

[tool.flit.sdist]
exclude = [
    "docs/",
    "tests/",
    "aiida_gui/frontend/node_modules/",
]


[tool.pylint.format]
max-line-length = 120
